#121 move tests out of tests package as only they were getting picked up by ci?
