+++ app/src/androidTest/java/org/kiwix/kiwixmobile/BasicTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.tests;
+package org.kiwix.kiwixmobile;
 
 import android.Manifest;
 import androidx.test.filters.LargeTest;
