+++ app/src/androidTest/java/org/kiwix/kiwixmobile/ZimTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.tests;
+package org.kiwix.kiwixmobile;
 
 //@LargeTest
 //@RunWith(AndroidJUnit4.class)
