+++ app/src/androidTest/java/org/kiwix/kiwixmobile/activities/package-info.java
@@ -15,4 +15,4 @@
  * Activity)
  */
 
-package org.kiwix.kiwixmobile.tests.activities;
+package org.kiwix.kiwixmobile.activities;
